pyproject.toml
[tool.poetry]
name = "retail-backend"
version = "0.1.0"
description = "A retail backend service with MySQL and S3 integration"
authors = ["Your Name brendancopp1@gmail.com"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.1"
uvicorn = "^0.15.0"
mysql-connector-python = "^8.0.26"
boto3 = "^1.26.137"
pydantic = "^1.8.2"
python-multipart = "^0.0.5"
requests = "^2.31.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
flake8 = "^6.1.0"
mypy = "^1.5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 